Add support for `hg add` and `hg status`.
Summary:
With the Hg Eden extension installed for an Eden mount, it is now possible to
create files and `hg add` them. When `hg status` is run, untracked (`?`) and
added (`A`) files will be listed. If an added file is deleted, it will be
correctly reported as missing (`!`) by `hg status`. Similarly, modified files
will also be included in the output of `hg status`.

The next step is to add support for `hg remove` to ensure removed files (`R`)
are also listed correctly by `hg status`. Support for ignored files also needs
to be added.

Reviewed By: simpkins

Differential Revision: D4223704

fbshipit-source-id: 49ff837bef8639031770c313979d98cad5e4b3fb