Override localrepo.invalidatedirstate to invalidate the dirstate.
Summary:
In Mercurial's `localrepo.py`, when `wlock()` is invoked (which locks the
non-store parts of the repository), it calls `invalidatedirstate()`, but not
`invalidate()`. When the lock is acquired, we want to be sure the `p1` in
`edendirstate` is up-to-date, so we need to make sure `edendirstate.invalidate()`
is called.

Note that the default implementation of `localrepo.invalidatedirstate()` is:

```
if hasunfilteredcache(self, 'dirstate'):
    for k in self.dirstate._filecache:
        try:
            delattr(self.dirstate, k)
        except AttributeError:
            pass
    delattr(self.unfiltered(), 'dirstate')
```

So despite its name, `edendirstate.invalidate()` is not called. Since our
`_filecache` property in `edendirstate` is a dummy value anyway, we don't lose
anything by dropping the default implementation. More importantly, the default
implementation provides no signal to `edendirstate` that it is being called
(short of adding dummy info to `_filecache`), so replacing it with our own
implementation is the best way to ensure we are notified.

Reviewed By: simpkins

Differential Revision: D4242129

fbshipit-source-id: 64ae156de50cf27bdc74ea0ef7ca87433bea85ed