use thrift directly inside the mercurial extension
Summary:
Use the eden.thrift helper library to create a thrift client, and use it
instead of the LameThriftClient module which executes the standalone pyremote
script.

If the normal thrift client extension cannot be imported this does fall back to
using the LameThriftClient code.  This happens if you are not using a build of
python that is compatibile with the compiled eden extension code.  I have left
this in for now to ease development until a compatible build of mercurial is
rolled out everywhere.

Reviewed By: bolinfest

Differential Revision: D3942717

fbshipit-source-id: a1ac8d731e3c96e7ad1bca2b6f5fe9d7c7c9ec41