change the "hg update" output to not report updated or removed counts
Summary:
One of the primary goals of eden is to avoid having to process the entire
repository state on update operations.  Therefore we do not build an accurate
count of the number of files updated or removed when switching between two
commits.

This overrides the function used to display the statistics after an update
operation so that we only report the number of conflicts, and not the number of
files affected.

Reviewed By: wez

Differential Revision: D4584244

fbshipit-source-id: f912d119929abd9a30f50740cbde47cf8410191f