import mercurial datapack, treemanifest sources
Summary:
This is a pre-cursor to having eden consume the data from
mercurial datapacks directly.

Some things to note:

* This is a mirror of a subset of the GPL code from the facebook-hg-rpms/fb-hgext dir
* This is not a fork!
* Nothing depends on this yet

I had to make a few minor edits to the code to make it pass our more strict
compilation flags; D4879285 is open to submit them back to the source of truth.

Reviewed By: DurhamG

Differential Revision: D4829860

fbshipit-source-id: fdbe612bda51aaacfe1f7faa1e2a0ec2ff9e04d6