don't use an edendirstate in non-eden repositories
Summary:
This fixes the eden hg extension to correctly return a normal dirstate object
if the repository is not inside an eden client mount point.

When using the share extension mercurial creates repo objects for both the
original working directory and the source repository.  We currently normally
configure eden hg clients to point at a shared repository that is not inside an
eden mount.  Recent versions of mercurial now seem to create and access
dirstate objects for the source repository even just to do basic operations
inside the shared repository.

Reviewed By: bolinfest

Differential Revision: D4882332

fbshipit-source-id: 4a702d3624c531b86002ad5e95e22192d26182e3