add --stdin option to pyremote
Summary:
We're trying and failing at passing binary strings through to pyremote
via subprocess.Popen when the string has an embedded NUL byte.

This is a quick workaround for this issue; passing the `--stdin` flag causes
pyremote to assume that stdin is a pickled list containing the function name
and arguments and is used in place of the normal handling for the trailing
arguments in the argument vector.

Reviewed By: simpkins

Differential Revision: D4872896

fbshipit-source-id: b86a33546d7076000acf6bf493f704b47a1b84d3