treemanifest: fix heap corruption in ConstantStringRef
Summary:
This memory is allocated using `operator new[]` and
must therefore be deleted with `delete []` to avoid breaking
the world.

I'll look at tidying up this code in the upstream hg extensions repo
to see if we can use `shared_ptr<string>` here today and either submit
that diff or this one there.

Just putting this up now to surface the issue.

Reviewed By: simpkins

Differential Revision: D4929693

fbshipit-source-id: 094ffa6b6fc57e67d7185c8968e69de6ff170f8a