implement dirstate.matches(), dirstate.walk() to enable `hg grep`.
Summary:
Relatively simplistic implementation of the matches and walk methods.

The approach is simple in concept: translate a mercurial matcher object into
a set of glob expressions that we can ask eden to evaluate to give us a list
of matching files.

For walks/matches that need to exclude based on ignored and unknown status,
we then issue a status call and amend the list of files.

For walk, we then need to return stat() style information for the files,
so we issue a call to eden to get that data.

Reviewed By: simpkins

Differential Revision: D4908820

fbshipit-source-id: b8eb53a028a4719244431747bbaa2ab887bf3718