Clean up what I think is a bad merge between D4908820 and D4935943.
Summary:
I am getting errors in some changes built on top of D4908820 because
`subprocess.Popen()` is getting non-string arguments for the command.
This appears to be because we have some cases of args being passed to
`LameThriftClient._call()` as a list instead of individually.

Also, `scmMarkCommitted()` was removed in D4935943, so this removes it again.

Reviewed By: wez

Differential Revision: D4984927

fbshipit-source-id: 874155d55748d8ec55126d614d905874e0c60102