adjust matcher and dirstate code for upstream mercurial changes
Summary:
The matcher module got refactored to be a bit more sane.
We may be able to re-do our layer over the top a bit more cleanly
with the new model, but for now, this diff just adjusts the wrapping;
we used to wrap the `match.__init__` method, but since `match` is now
a simple function, we need to tweak the wrapping a little bit.

In addition, the `parsers` module used by dirstate is now accessed
via a `policy` accessor, so adjust to use that.

Reviewed By: simpkins

Differential Revision: D5185002

fbshipit-source-id: be4958ea3bef2065a131bc251762038fdf09477a