deal with API changes in the mercurial dirstate constructor
Summary:
Upstream mercurial has recently changed the number of arguments passed to the
dirstate constructor.

If mercurial has the sparse module, also pass a sparsematchfn argument to the
dirstate constructor.  Otherwise assume it is an older version of mercurial
that does not require this argument.

There are 77 mercurial commits between when the sparse module was first added
(ba5d897) and when the extra dirstate parameter was added (fb32039).  This code
won't work when using a version of mercurial from between those too commits,
but that probably shouldn't be an issue in practice.

Reviewed By: DurhamG

Differential Revision: D5422335

fbshipit-source-id: 4af748e9fd683fd4711fee61eeb7aefaf76cd007