no more lame thrift on devservers
Summary:
The LameThrift client has a pretty terrible overhead for
operations such as rebase that need to make a moderate to large
number of thrift calls.

In my rebase testing scenario is is responsible for 50% of the
runtime of the rebase.

The thrift compiler generated code depends on python 2.7.6 because that is the
first version of python to support unicode literal strings being passed as the
`fmt` argument of `struct.pack` and `struct.unpack`.

Our Linux systems are running 2.7.5.  This diff monkey patches the pack
functions to decode the unicode format strings if we're running on 2.7.5.

Reviewed By: mjpieters

Differential Revision: D5929978

fbshipit-source-id: 71424e59e382e3746d01ed90616a0b0e029371d9