Move _copymap into eden_dirstate_map to reflect changes to Mercurial upstream.
Summary:
DurhamG is in the process of refactoring the `dirstate` upstream.
https://phab.mercurial-scm.org/D759 is the top of the stack that has already
landed and there is more coming in the https://phab.mercurial-scm.org/D983
stack.

In particular, https://phab.mercurial-scm.org/D756 moved the `_copymap` field,
so this change updates `eden_dirstate_map` to mirror that change, which is nice
because it makes our `eden_dirstate.py` file a bit smaller. The more we can rely
on delegation (to `dirstatemap`/`eden_dirstate_map`) rather than inheritance
(of `dirstate` by `eden_dirstate`), the more resilient we should be to upstream
changes in the long run.

Reviewed By: wez

Differential Revision: D6005469

fbshipit-source-id: 79f36abd3e13e93f779b770f5f37edf7bb8d47b6