Implement eden_dirstate.write() so it updates the .hg/dirstate file.
Summary:
It turns out that Nuclide uses modifications to `.hg/dirstate` as the signal
to invalidate its internal Hg state caches. Previously, Eden would never
create an `.hg/dirstate` file, so now we do so in order to trigger the necessary
updates in Nuclide.

The implementation of `eden_dirstate.write()` is a port of `dirstate.write()`
with some logic removed because it is unnecessary for Eden. As nothing should be
reading the contents of the `.hg/dirstate` file, we write the hex values of the
parents there, just so we have something useful.

For reference, here is where Nuclide creates the Watchman subscription:

https://github.com/facebook/nuclide/blob/release-v0.262.0/pkg/nuclide-hg-rpc/lib/HgService.js#L593-L602

and here is where it adds a subscriber that dispatches an "Hg repo state
changed" event:

https://github.com/facebook/nuclide/blob/release-v0.262.0/pkg/nuclide-hg-rpc/lib/HgService.js#L667

Reviewed By: wez

Differential Revision: D6116105

fbshipit-source-id: 2d1785dac9fcbc3ee42e5bff97e6a69168090c3d