fix sys.path to allow importing of the eden.dirstate module
Summary:
D6179950 updated the eden hg extension to import the eden.dirstate module which
is built as a separate python library, and installed outside of the
hgext3rd/eden package.  However, it failed trying to import it since it did not
modify sys.path to include the correct directory.

Previously EdenThriftClient was modifying sys.path so it could import our
thrift modules, but it only did this after trying to import eden.dirstate.

Reviewed By: bolinfest

Differential Revision: D6263604

fbshipit-source-id: 3de329f038ff37405e4ac9973813447e09975928